Add Abstract Kex class
