+++ lib/net/ssh/transport/kex/diffie_hellman_group1_sha1.rb
@@ -1,8 +1,4 @@
-require 'net/ssh/buffer'
-require 'net/ssh/errors'
-require 'net/ssh/loggable'
-require 'net/ssh/transport/openssl'
-require 'net/ssh/transport/constants'
+require 'net/ssh/transport/kex/abstract'
 
 module Net
   module SSH
@@ -10,10 +6,7 @@ module Net
       module Kex
         # A key-exchange service implementing the "diffie-hellman-group1-sha1"
         # key-exchange algorithm.
-        class DiffieHellmanGroup1SHA1
-          include Loggable
-          include Constants
-
+        class DiffieHellmanGroup1SHA1 < Abstract
           # The value of 'P', as a string, in hexadecimal
           P_s = "FFFFFFFF" "FFFFFFFF" "C90FDAA2" "2168C234" +
                 "C4C6628B" "80DC1CD1" "29024E08" "8A67CC74" +
@@ -29,47 +22,6 @@ module Net
 
           # The group constant
           G = 2
-
-          attr_reader :algorithms
-          attr_reader :connection
-          attr_reader :data
-          attr_reader :dh
-
-          # Create a new instance of the DiffieHellmanGroup1SHA1 algorithm.
-          # The data is a Hash of symbols representing information
-          # required by this algorithm, which was acquired during earlier
-          # processing.
-          def initialize(algorithms, connection, data)
-            @algorithms = algorithms
-            @connection = connection
-
-            @data = data.dup
-            @dh = generate_key
-            @logger = @data.delete(:logger)
-          end
-
-          # Perform the key-exchange for the given session, with the given
-          # data. This method will return a hash consisting of the
-          # following keys:
-          #
-          # * :session_id
-          # * :server_key
-          # * :shared_secret
-          # * :hashing_algorithm
-          #
-          # The caller is expected to be able to understand how to use these
-          # deliverables.
-          def exchange_keys
-            result = send_kexinit
-            verify_server_key(result[:server_key])
-            session_id = verify_signature(result)
-            confirm_newkeys
-
-            return { session_id: session_id,
-                     server_key: result[:server_key],
-                     shared_secret: result[:shared_secret],
-                     hashing_algorithm: digester }
-          end
 
           def digester
             OpenSSL::Digest::SHA1
@@ -162,57 +114,6 @@ module Net
             result[:server_sig] = sig_buffer.read_string
 
             return result
-          end
-
-          # Verify that the given key is of the expected type, and that it
-          # really is the key for the session's host. Raise Net::SSH::Exception
-          # if it is not.
-          def verify_server_key(key) #:nodoc:
-            if key.ssh_type != algorithms.host_key
-              raise Net::SSH::Exception,
-                "host key algorithm mismatch " +
-                "'#{key.ssh_type}' != '#{algorithms.host_key}'"
-            end
-
-            blob, fingerprint = generate_key_fingerprint(key)
-
-            raise Net::SSH::Exception, "host key verification failed" unless connection.host_key_verifier.verify(key: key, key_blob: blob, fingerprint: fingerprint, session: connection)
-          end
-
-          def generate_key_fingerprint(key)
-            blob = Net::SSH::Buffer.from(:key, key).to_s
-
-            fingerprint = Net::SSH::Authentication::PubKeyFingerprint.fingerprint(blob, @connection.options[:fingerprint_hash] || 'SHA256')
-
-            [blob, fingerprint]
-          rescue ::Exception => e
-            [nil, "(could not generate fingerprint: #{e.message})"]
-          end
-
-          # Verify the signature that was received. Raise Net::SSH::Exception
-          # if the signature could not be verified. Otherwise, return the new
-          # session-id.
-          def verify_signature(result) #:nodoc:
-            response = build_signature_buffer(result)
-
-            hash = digester.digest(response.to_s)
-
-            raise Net::SSH::Exception, "could not verify server signature" unless connection.host_key_verifier.verify_signature { result[:server_key].ssh_do_verify(result[:server_sig], hash) }
-
-            return hash
-          end
-
-          # Send the NEWKEYS message, and expect the NEWKEYS message in
-          # reply.
-          def confirm_newkeys #:nodoc:
-            # send own NEWKEYS message first (the wodSSHServer won't send first)
-            response = Net::SSH::Buffer.new
-            response.write_byte(NEWKEYS)
-            connection.send_message(response)
-
-            # wait for the server's NEWKEYS message
-            buffer = connection.next_message
-            raise Net::SSH::Exception, "expected NEWKEYS" unless buffer.type == NEWKEYS
           end
         end
       end
